requirements.txt
anyio==3.6.1
asyncpg==0.26.0
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
fastapi==0.82.0
greenlet==1.1.3
h11==0.13.0
httptools==0.4.0
idna==3.3
install==1.3.5
passlib==1.7.4
pycparser==2.21
pydantic==1.10.1
PyJWT==2.4.0
python-dotenv==0.21.0
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==1.4.40
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.3
watchfiles==0.16.1
websockets==10.3
